zmy_pip_requirements.txt
absl-py==2.1.0
addict==2.4.0
ansicon==1.89.0
blessed==1.20.0
cachetools==5.3.3
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colour==0.1.5
colour-demosaicing==0.2.5
colour-science==0.4.4
contourpy==1.2.1rc1
cycler==0.12.1
decorator==4.4.2
docker-pycreds==0.4.0
einops==0.7.0rc2
filelock==3.9.0
fonttools==4.53.1
fsspec==2023.4.0
fvcore==0.1.5.post20221221
gitdb==4.0.9
GitPython==3.1.9
google-auth==2.4.0
google-auth-oauthlib==0.4.1
grpcio==1.66.0rc5
h5py==3.9.0
huggingface-hub==0.9.1
idna==3.4
imageio==2.34.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.4.3
iopath==0.1.9
Jinja2==3.1.2
joblib==1.0.1
jupter==0.1.0
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.7
MarkupSafe==2.1.3
matplotlib==3.7.4
moviepy==1.0.3
mpmath==1.3.0
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==24.0
pandas==2.2.2
Pillow==9.3.0
platformdirs==4.2.2
portalocker==2.8.2
proglog==0.1.10
protobuf==3.19.2
psutil==5.9.8
py3nvml==0.2.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0b1
requests==2.28.1
requests-oauthlib==1.4.1
rsa==4.9
safetensors==0.4.4rc0
scikit-image==0.24.0rc1
scikit-learn==1.1.0
scipy==1.11.4
seaborn==0.13.2
sentry-sdk==2.9.0
setproctitle==1.3.3
six==1.9.0
smmap==5.0.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.0.0
tifffile==2024.7.21
timm==1.0.9
torch==2.1.2
torch-ema==0.3
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.4.0
tzdata==2024.1
urllib3==1.26.13
wandb==0.17.7
wcwidth==0.2.9
Werkzeug==3.0.4
xmltodict==0.14.2
yacs==0.1.8
zipp==3.9.1